avro/tests: extract test data generator
This will be useful for generating data in other tests.
